Делаем запросы для экрана пользователя, получаем URL аватарки пользователя, замена на "новое" API для KVO, переделка сетевого запроса на дженериках, показ алерта с ошибкой
