+++ ImageFeed/Auth/AuthViewController.swift
@@ -40,6 +40,6 @@ extension AuthViewController: WebViewViewControllerDelegate {
     
     func webViewViewController(_ vc: WebViewViewController, didAuthenticateWithCode code: String) {
         delegate?.authViewController(self, didAuthenticateWithCode: code)
-        UIBlockingProgressHUD.show()
+ //       UIBlockingProgressHUD.show()
     }
 }
