+++ ImageFeed/Auth/OAuth2Service.swift
@@ -59,7 +59,7 @@ extension OAuth2Service {
         for request: URLRequest,
         completion: @escaping (Result<OAuthTokenResponseBody, Error>) -> Void) -> URLSessionTask {
             let decoder = JSONDecoder()
-            return urlSession.data(for: request) { (result: Result<Data, Error>) in
+            return urlSession.objectTask(for: request) { (result: Result<Data, Error>) in
                 let response = result.flatMap {data -> Result<OAuthTokenResponseBody, Error> in
                     Result { try decoder.decode(OAuthTokenResponseBody.self, from: data)}
                 }
@@ -108,32 +108,3 @@ extension URLRequest {
 }
 
 // MARK: - Network Connection
-extension URLSession {
-    func data(
-        for request: URLRequest,
-        completion: @escaping (Result<Data, Error>) -> Void) -> URLSessionTask {
-            let fulfillCompletion: (Result<Data, Error>) -> Void = { result in
-                DispatchQueue.main.async {
-                    completion(result)
-                }
-            }
-            let task = dataTask(with: request, completionHandler: {data, response, error in
-                if let data = data,
-                   let response = response,
-                   let statusCode = (response as? HTTPURLResponse)?.statusCode
-                {
-                    if 200 ..< 300 ~= statusCode {
-                        fulfillCompletion(.success(data))
-                    } else {
-                        fulfillCompletion(.failure(NetworkError.httpStatusCode(statusCode)))
-                    }
-                } else if let error = error {
-                    fulfillCompletion(.failure(NetworkError.urlRequestError(error)))
-                } else {
-                    fulfillCompletion(.failure(NetworkError.urlSessionError))
-                }
-            })
-            task.resume()
-            return task
-        }
-}
